// src/app/(DashboardLayout)/components/types/supabase.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export interface Database {
  public: {
    Tables: {
      config: {
        Row: {
          created_at: string | null
          id: number
          modified_at: string | null
          name: string | null
          status: number | null
          user_id: string | null
          values: Json | null
          version: number | null
        }
        Insert: {
          created_at?: string | null
          id?: number
          modified_at?: string | null
          name?: string | null
          status?: number | null
          user_id?: string | null
          values?: Json | null
          version?: number | null
        }
        Update: {
          created_at?: string | null
          id?: number
          modified_at?: string | null
          name?: string | null
          status?: number | null
          user_id?: string | null
          values?: Json | null
          version?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "config_user_id_fkey"
            columns: ["user_id"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
      users: {
        Row: {
          active: boolean | null
          code: string | null
          created_at: string | null
          email: string | null
          fullname: string | null
          id: number
          last_login: string | null
          status: string | null
          subscription: string | null
          user_id: string | null
        }
        Insert: {
          active?: boolean | null
          code?: string | null
          created_at?: string | null
          email?: string | null
          fullname?: string | null
          id?: number
          last_login?: string | null
          status?: string | null
          subscription?: string | null
          user_id?: string | null
        }
        Update: {
          active?: boolean | null
          code?: string | null
          created_at?: string | null
          email?: string | null
          fullname?: string | null
          id?: number
          last_login?: string | null
          status?: string | null
          subscription?: string | null
          user_id?: string | null
        }
        Relationships: [
          {
            foreignKeyName: "users_user_id_fkey"
            columns: ["user_id"]
            referencedRelation: "users"
            referencedColumns: ["id"]
          }
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}